Remove old php-cs-fixer1 implementation + add configurable priority on task config
