+++ src/Task/PhpCsFixer.php
@@ -4,33 +4,45 @@ declare(strict_types=1);
 
 namespace GrumPHP\Task;
 
+use GrumPHP\Formatter\PhpCsFixerFormatter;
 use GrumPHP\Runner\TaskResult;
 use GrumPHP\Runner\TaskResultInterface;
 use GrumPHP\Task\Context\ContextInterface;
+use GrumPHP\Task\Context\GitPreCommitContext;
 use GrumPHP\Task\Context\RunContext;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-/**
- * Php-cs-fixer task.
- */
-class PhpCsFixer extends AbstractPhpCsFixerTask
+class PhpCsFixer extends AbstractExternalTask
 {
+    /**
+     * @var PhpCsFixerFormatter
+     */
+    protected $formatter;
+
     public static function getConfigurableOptions(): OptionsResolver
     {
         $resolver = new OptionsResolver();
         $resolver->setDefaults([
+            'allow_risky' => null,
+            'cache_file' => null,
             'config' => null,
-            'config_file' => null,
-            'fixers' => [],
-            'level' => null,
+            'rules' => [],
+            'using_cache' => null,
+            'config_contains_finder' => true,
             'verbose' => true,
+            'diff' => false,
+            'triggered_by' => ['php'],
         ]);
 
+        $resolver->addAllowedTypes('allow_risky', ['null', 'bool']);
+        $resolver->addAllowedTypes('cache_file', ['null', 'string']);
         $resolver->addAllowedTypes('config', ['null', 'string']);
-        $resolver->addAllowedTypes('config_file', ['null', 'string']);
-        $resolver->addAllowedTypes('fixers', ['array']);
-        $resolver->addAllowedTypes('level', ['null', 'string']);
+        $resolver->addAllowedTypes('rules', ['array']);
+        $resolver->addAllowedTypes('using_cache', ['null', 'bool']);
+        $resolver->addAllowedTypes('config_contains_finder', ['bool']);
         $resolver->addAllowedTypes('verbose', ['bool']);
+        $resolver->addAllowedTypes('diff', ['bool']);
+        $resolver->addAllowedTypes('triggered_by', ['array']);
 
         return $resolver;
     }
@@ -38,30 +50,62 @@ class PhpCsFixer extends AbstractPhpCsFixerTask
     /**
      * {@inheritdoc}
      */
+    public function canRunInContext(ContextInterface $context): bool
+    {
+        return $context instanceof GitPreCommitContext || $context instanceof RunContext;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
     public function run(ContextInterface $context): TaskResultInterface
     {
-        $files = $context->getFiles()->name('*.php');
+        $config = $this->getConfig()->getOptions();
+        $files = $context->getFiles()->extensions($config['triggered_by']);
         if (0 === \count($files)) {
             return TaskResult::createSkipped($this, $context);
         }
 
-        $config = $this->getConfig()->getOptions();
         $this->formatter->resetCounter();
 
         $arguments = $this->processBuilder->createArgumentsForCommand('php-cs-fixer');
         $arguments->add('--format=json');
         $arguments->add('--dry-run');
-        $arguments->addOptionalArgument('--level=%s', $config['level']);
+        $arguments->addOptionalBooleanArgument('--allow-risky=%s', $config['allow_risky'], 'yes', 'no');
+        $arguments->addOptionalArgument('--cache-file=%s', $config['cache_file']);
         $arguments->addOptionalArgument('--config=%s', $config['config']);
-        $arguments->addOptionalArgument('--config-file=%s', $config['config_file']);
-        $arguments->addOptionalArgument('--verbose', $config['verbose']);
-        $arguments->addOptionalCommaSeparatedArgument('--fixers=%s', $config['fixers']);
-        $arguments->add('fix');
 
-        if ($context instanceof RunContext && null !== $config['config_file']) {
-            return $this->runOnAllFiles($context, $arguments);
+        if ($rules = $config['rules']) {
+            $arguments->add(sprintf(
+                '--rules=%s',
+                // Comma-delimit rules if specified as a list; otherwise JSON-encode.
+                array_values($rules) === $rules ? implode(',', $rules) : json_encode($rules)
+            ));
         }
 
-        return $this->runOnChangedFiles($context, $arguments, $files);
+        $canUseIntersection = !($context instanceof RunContext) && $config['config_contains_finder'];
+
+        $arguments->addOptionalBooleanArgument('--using-cache=%s', $config['using_cache'], 'yes', 'no');
+        $arguments->addOptionalArgument('--path-mode=intersection', $canUseIntersection);
+        $arguments->addOptionalArgument('--verbose', $config['verbose']);
+        $arguments->addOptionalArgument('--diff', $config['diff']);
+        $arguments->add('fix');
+
+        if ($context instanceof GitPreCommitContext || !$config['config_contains_finder']) {
+            $arguments->addFiles($files);
+        }
+
+        $process = $this->processBuilder->buildProcess($arguments);
+        $process->run();
+
+        if (!$process->isSuccessful()) {
+            $messages = [$this->formatter->format($process)];
+            $suggestions = [$this->formatter->formatSuggestion($process)];
+            $errorMessage = $this->formatter->formatErrorMessage($messages, $suggestions);
+
+            return TaskResult::createFailed($this, $context, $errorMessage);
+        }
+
+        return TaskResult::createPassed($this, $context);
     }
 }
